+++ frontend-umi/src/pages/user/index.tsx
@@ -1,0 +1,3 @@
+import AccountCenterPage from '@/pages/account/center';
+
+export default AccountCenterPage;